rename updater2 -> updater, updater -> legacyupd
