+++ internal/legacyupd/check.go
@@ -1,4 +1,4 @@
-package updater
+package legacyupd
 
 import (
 	"context"
