+++ internal/legacyupd/check_test.go
@@ -1,4 +1,4 @@
-package updater
+package legacyupd
 
 import (
 	"os"
