+++ internal/legacyupd/hide_lin_mac.go
@@ -14,7 +14,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package updater
+package legacyupd
 
 func hideFile(path string) error {
 	return nil
