+++ internal/legacyupd/hide_win.go
@@ -1,6 +1,6 @@
 // +build windows
 
-package updater
+package legacyupd
 
 import (
 	"errors"
