+++ internal/legacyupd/requester.go
@@ -1,4 +1,4 @@
-package updater
+package legacyupd
 
 import (
 	"context"
