+++ internal/updater/checker.go
@@ -1,4 +1,4 @@
-package updater2
+package updater
 
 import (
 	"encoding/json"
