+++ internal/updater/fetcher_windows.go
@@ -1,6 +1,6 @@
 // +build windows
 
-package updater2
+package updater
 
 import "github.com/ActiveState/cli/internal/unarchiver"
